Fix: createToast amount 타입 변경

+++ MeokPT/Sources/Diet/AddFoodFeature.swift
@@ -65,7 +65,7 @@ struct AddFoodFeature {
     enum DelegateAction: Equatable {
         case dismissSheet
         case addFoodToDiet(foodName: String, amount: Double, calories: Double, carbohydrates: Double, protein: Double, fat: Double, dietaryFiber: Double, sugar: Double, sodium: Double)
-        case createToast(foodName: String, amount: Int)
+        case createToast(foodName: String, amount: Double)
     }
     
     var body: some ReducerOf<Self> {
